#4 Refactor: 와인 뉴스 기능 설정 - 와인 뉴스 validation 처리 - 와인 뉴스 오류 메시지 WINE_NEWS_NOT_FOUDN로 변경 - 와인 뉴스 delete 예외 처리

+++ src/main/java/com/drinkeg/drinkeg/service/wineNewsService/WineNewsServiceImpl.java
@@ -30,7 +30,7 @@ public class WineNewsServiceImpl implements WineNewsService {
     public WineNewsResponseDTO getWineNewsById(Long WineNewsId) {
         return WineNewsConverter.toWineNewsResponseDTO(
                 wineNewsRepository.findById(WineNewsId)
-                        .orElseThrow(() -> new GeneralException(ErrorStatus.WINENEWS_NOT_FOUND)));
+                        .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_NEWS_NOT_FOUND)));
     }
 
     @Override
@@ -44,13 +44,15 @@ public class WineNewsServiceImpl implements WineNewsService {
     public WineNewsResponseDTO updateWineNews(Long WineNewsId, WineNewsRequestDTO WineNewsRequestDTO) {
         return WineNewsConverter.toWineNewsResponseDTO(
                 wineNewsRepository.findById(WineNewsId)
-                        .orElseThrow(() -> new GeneralException(ErrorStatus.WINENEWS_NOT_FOUND))
+                        .orElseThrow(() -> new GeneralException(ErrorStatus.WINE_NEWS_NOT_FOUND))
                         .updateTitle(WineNewsRequestDTO.getTitle())
                         .updateCardNewsImg(WineNewsRequestDTO.getCardNewsImg()));
     }
 
     @Override
     public void deleteWineNews(Long WineNewsId) {
+        if (!wineNewsRepository.existsById(WineNewsId))
+            throw new GeneralException(ErrorStatus.WINE_NEWS_NOT_FOUND);
         wineNewsRepository.deleteById(WineNewsId);
     }
 }
